Create models for polymorphic content, create custom model fields
Create a Content model that represents the modules' contents and
define a generic relations to associate any object with the content
object.
Create an abstract ItemBase model that provides the common fields
for all Content models. Then create Text, File, Image, Video models
that inherit from the ItemBase abstract model and contain their own
unique fields.
Create a new file named fields.py inside the courses application and
then create a custom model field named OrderField that inherits from
the PositiveIntegerField provided by Django and calculates the order
of a model instance and assings it to it. Add this new field to
the Content and Module models.