Create class-based views to manage courses
Create the OwnerMixin and OwnerEditMixin mixins. Use these mixins
together with the ListView, CreateView, UpdateView, DeleteView views
to create the following views:
1.ManageCourseListView that lists the courses created by the user.
2.CourseCreateView that uses a model form to create a new Course
object
3.CourseUpdateView that allows the editing of an existing Course
object
4. CourseDeleteView that allows to delete Course objects.
Create a group named Instructors to allow the users to create,
change, delete and view content objects.
Restrict access to the views so that only users with appropriate
permissions can add, change or delete Courses objects.
Create URLs in the 'courses/urls.py' file and templates for the views.