Create views to manage course modules and their contents
Create the CourseModuleUpdateView view that handle the formset to
add, update and delete modules for a specific course. This view
inherits from the TemplateResponseMixin mixin and the basic View view.
Configure a URL pattern for this view and create a template.
Create the ContentCreateUpdateView view that allows to create and
update different models' contents. Configure URl patterns and create
a template for this view.
Create the ContentDeleteViev view for deleting contents.
Create the ModuleContentListView view that gets the Module object
with the given ID that belongs to the current user and renders a
template with the given module. Also create a template and configure
a URL pattern for the view.