Add functionality to reorder modules and their contents
Using the CsrfExemptMixin and JsonRequestResponseMixin from the
django-braces library, create the ModelOrderView view that receives
the new order of module IDs encoded in JSON and updates the order
accordingly. Build a URL pattern for this view.
Create similar view for contents ordering.
Add JavaScript code to the content_list.html template, that provides
the order change functionality.