Create views to display courses
Create the CourseListView view that inherits from TemplateResponseView
and View classes. This view is responsible for displaying a list
of courses on the main page. Add a URL pattern for this view to the
main urls.py file. Create a template for the view.
Create the DetailCourseView view that inherits from the django
DetailView and displays a single course overview. Add a URL pattern
to the courses app urls.py file and create a template for it.