Build functionality for students
Create an app called students in the project. Create the
StudentRegistrationView view to allows students to register on the
site. This view inherits from the django generic CreateView.
Add a URL pattern and create a template for the view.
Edit the models.py of the course application and add a new
ManyToMany field to the Course model named students, that represents
students enrolled on a course.
Create the StudentsEnrollCourseView that inherits from the FormView
and handles students enrolling on courses.
Create the StudentCourseListView and StudentCourseDetailView views
that inherit from the django generic ListView and DetailView
respectively. Add URL patterns for these views and create templates.