Create a chat application
Create a new app called chat. Implement the course_chat_room view
that renders a template to display  the course's chat. Add a URL
pattern and create the 'chat/room.html' template.
Install the Channels library and set up the required basic ASGI
application routing for it to manage HTTP requests.
Create a python file 'consumers.py' and write a consumer, which is
the equivalent of Django views for asynchronous applications, to
handle connecting to a WebSocket, receiving and sending messages
and disconnecting. Define a URL to route connections to the
ChatConsumer consumer. Add a new route for the websocket protocol
in the 'educa/asgi.py' file. Using JavaScript, Build a WebSocket
client to establish a  connection with the WebSocket in the course
chat room template and be able to send/receive messages.
Set up a channel layer with Redis to distribute messages between
consumers.